Revert "chore[#33]: reorganize Professor pages into modules"
