+++ client/src/pages/dashboard/DashboardPage.tsx
@@ -1,6 +1,6 @@
 import PageTemplate from "../../components/PageTemplate";
 import { formatTodayEs } from "../../utils/date";
-import ProfessorHome from "../professor/dashboard/ProfessorHome";
+import ProfessorHome from "./ProfessorHome";
 import StudentHome from "./StudentHome";
 import { useUserStore } from "../../store/userStore";
