+++ client/src/pages/exams/ExamCreatePage.tsx
@@ -1,20 +1,19 @@
 import { useRef, useState } from 'react';
 import type { CSSProperties } from 'react';
-import '../../../components/exams/ExamForm.css';
-import '../../../components/shared/Toast.css';
-import { ExamForm } from '../../../components/exams/ExamForm';
-import type { ExamFormHandle } from '../../../components/exams/ExamForm';
-import { Toast, useToast } from '../../../components/shared/Toast';
-import { readJSON } from '../../../services/storage/localStorage';
-import PageTemplate from '../../../components/PageTemplate';
-import GlobalScrollbar from '../../../components/GlobalScrollbar';
+import '../../components/exams/ExamForm.css';
+import '../../components/shared/Toast.css';
+import { ExamForm } from '../../components/exams/ExamForm';
+import type { ExamFormHandle } from '../../components/exams/ExamForm';
+import { Toast, useToast } from '../../components/shared/Toast';
+import { readJSON } from '../../services/storage/localStorage';
+import PageTemplate from '../../components/PageTemplate';
+import GlobalScrollbar from '../../components/GlobalScrollbar';
 import './ExamCreatePage.css';
-import { generateQuestions, createExamApproved, type GeneratedQuestion } from '../../../services/exams.service';
-import AiResults from '../../exams/AiResults';
-import { normalizeToQuestions, cloneQuestion, replaceQuestion, reorderQuestions } from '../../exams/ai-utils';
-import { isValidGeneratedQuestion } from '../../../utils/aiValidation';
+import { generateQuestions, createExamApproved, type GeneratedQuestion } from '../../services/exams.service';
+import AiResults from './AiResults';
+import { normalizeToQuestions, cloneQuestion, replaceQuestion, reorderQuestions } from './ai-utils';
+import { isValidGeneratedQuestion } from '../../utils/aiValidation';
 import { useNavigate, useSearchParams } from 'react-router-dom';
-
 
 
 const layoutStyle: CSSProperties = {
